ABM carreras lista, modificacion Nav
